Extract DCT specifications from YAML fragments

This function takes a file containing YAML fragments and extracts the DCT specifications from them.

example_dct_spec

extract_dct_specs(file, text, delimiterRegEx = "^---$",
  dctContainer = "dct", ignoreOddDelimiters = FALSE, silent = FALSE)

Arguments

file
The file containing the YAML fragments.

text
A character vector containing the YAML fragments, where every element should represent one line in the file.

delimiterRegEx
The regular expression used to locate YAML fragments

dctContainer
The container of the DCT specifications in the YAML fragments. Because only DCT specifications are read that are stored in this container, the files can contain YAML fragments with other data, too, without interfering with the parsing of the DCT specifications.

ignoreOddDelimiters
Whether to throw an error (FALSE) or delete the last delimiter (TRUE) if an odd number of delimiters is encountered.

silent
Whether to be silent (TRUE) or informative (FALSE).

Format

An object of class character of length 1.

Value

A list of DCT specifications in an object of class dtcRawSpecList.

Details

This function is called by extract_dct_dir(); it is normally not necessary to call this function directly.

Examples

extract_dct_specs(text=unlist(strsplit(dct::example_dct_spec, '\n')));

#> Error in eval(parse(text = text, keep.source = FALSE), envir): argument "file" is missing, with no default