Extract DCT specifications from YAML fragments

This function takes a file containing YAML fragments and extracts the DCT specifications from them.

example_dct_spec

extract_dct_specs(file, text, delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE, silent = FALSE)

Arguments

file
The file containing the YAML fragments.

text
A character vector containing the YAML fragments, where every element should represent one line in the file.

delimiterRegEx
The regular expression used to locate YAML fragments

ignoreOddDelimiters
Whether to throw an error (FALSE) or delete the last delimiter (TRUE) if an odd number of delimiters is encountered.

silent
Whether to be silent (TRUE) or informative (FALSE).

Format

An object of class character of length 1.

Value

A list of DCT specifications in an object of class dtcRawSpecList.

Details

This function is called by extract_dct_dir(); it is normally not necessary to call this function directly.

Examples

extract_dct_specs(text=unlist(strsplit(dct::example_dct_spec, '\n')));

#> [[1]]
#> $id
#> [1] "attitude_xl678lqgq"
#> 
#> $label
#> [1] "Altitude"
#> 
#> $date
#> [1] "2018-11-28"
#> 
#> $source
#> $source$label
#> [1] "Reasoned Action Approach"
#> 
#> $source$xdoi
#> [1] "ISBN"
#> 
#> $source$spec
#> [1] "p. 22"
#> 
#> 
#> $defs
#> $defs$def
#> [1] "Attitude Def"
#> 
#> $defs$source
#> $defs$source$spec
#> [1] "pp. 11-12"
#> 
#> 
#> 
#> $measure_dev
#> $measure_dev$instr
#> [1] "Use semantic differentials with root \"For me, [TARGET BEHAVIOR] is ...\" and a bidimensional scale where the right-most anchor expresses a beneficial or pleasant evaluation and the left-most anchor expresses the opposite undesirable/harmful/unpleasant evaluation."
#> 
#> 
#> $manipulate_dev
#> $manipulate_dev$instr
#> [1] "persuasive communication"
#> 
#> $manipulate_dev$source
#> $manipulate_dev$source$label
#> [1] "pers comm matrix"
#> 
#> $manipulate_dev$source$xdoi
#> [1] "doi of article"
#> 
#> 
#> 
#> $measure_code
#> $measure_code$instr
#> [1] "Operationalisations that measure beneficial or pleasant aspects of the latent disposition or tendency to respond favourably versus unfavourably to [target behavior], for example using the semantic differentials good/bad, pleasant/unpleasant, wise/unwise."
#> 
#> $measure_code$source
#> [1] "none"
#> 
#> 
#> $manipulate_code
#> $manipulate_code$instr
#> [1] "Code any presentation of a stimulus asserting that engaging in the target behavior will yield positive or negative consequences."
#> 
#> $manipulate_code$source
#> [1] "none"
#> 
#> 
#> $aspect_code
#> $aspect_code$instruction
#> [1] "Expressions of any consequences for oneself that one perceives to be caused by engaging in the target behavior."
#> 
#> $aspect_code$source
#> [1] "none"
#> 
#> 
#> $parent
#> $parent$id
#> [1] "intention_xl678lqgr"
#> 
#> 
#> $alters
#> $alters$id
#> NULL
#> 
#> 
#> $datetime
#> [1] 1543359600
#> 
#> attr(,"class")
#> [1] "dtcRawSpec"
#> 
#> attr(,"class")
#> [1] "dtcRawSpecList"